node-main/deps/v8/test/webkit/fast/js/Promise-resolve-with-then-reject-expected.txt

Test whether Promise treats thenable correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

The promise is not rejected now.
PASS value.then is called.
PASS thisValue is value
PASS rejected
PASS result is "hello"
PASS successfullyParsed is true

TEST COMPLETE
